move button to folder and update paths
