+++ frontend/src/app/accountCreation/MfaSelect/MfaSelect.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 
 import { Card } from "../../commonComponents/Card/Card";
 import { CardBackground } from "../../commonComponents/CardBackground/CardBackground";
-import Button from "../../commonComponents/Button";
+import Button from "../../commonComponents/Button/Button";
 import RadioGroup from "../../commonComponents/RadioGroup";
 import StepIndicator from "../../commonComponents/StepIndicator";
 import { accountCreationSteps } from "../../../config/constants";
