+++ frontend/src/app/accountCreation/SecurityQuestion/SecurityQuestion.tsx
@@ -4,7 +4,7 @@ import { Card } from "../../commonComponents/Card/Card";
 import { CardBackground } from "../../commonComponents/CardBackground/CardBackground";
 import Dropdown from "../../commonComponents/Dropdown";
 import TextInput from "../../commonComponents/TextInput";
-import Button from "../../commonComponents/Button";
+import Button from "../../commonComponents/Button/Button";
 import StepIndicator from "../../commonComponents/StepIndicator";
 import {
   accountCreationSteps,
