+++ frontend/src/app/testResults/TestResultInputForm.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import RadioGroup from "../commonComponents/RadioGroup";
-import Button from "../commonComponents/Button";
+import Button from "../commonComponents/Button/Button";
 import { COVID_RESULTS, TEST_RESULT_DESCRIPTIONS } from "../constants";
 import { TestResult } from "../testQueue/QueueItem";
